add el_new and el_destroy functions for stable ABI
el_new will allocate on heap "el" object, so user does not have
to know any internals of "struct el". This will make ABI stable
in case when anything in "struct el" changes.

el_oinit can still be used to allocate memory on stack, but with
it ABI may be broken, so it should be used only on systems when
user can afford to recompile all programs that use embedlog lib.

* include/embedlog.h.in
  add 2 new functions, el_init and el_destroy

* man/
  add description for 2 new functions, describe notes about
  stable ABI and stack allocated methods.

* tst/
  add test for 2 new functions

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>