configure.ac: remove unneeded AC_CHECK_FUNCS
This patch removes check for functions that we cannot live
without anyway. Like - it's pointles to check for strlen(),
if that function is not available, program will not compile
no matter what. But checks like access(), if these functions
are not available, we can use different one (usually slower),
and we can still work.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>