src/snprintf.c: fix compilation warning with -ansi
When -ansi flag is enabled, configure won't find snprintf() function
and will set NEED_SNPRINTF_ONLY. But with it "credits" variable
will not be present in compilation unit as it requries
PREFER_PORTABLE_SNPRINTF - which we don't want to set. So
"str_arg = credits;" should be used only when PREFER_PORTABLE_SNPRINTF
is defined.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>