src/el-decode-number.c: fix potential loose of precision on 64bit systems
*out |= (n[i] & 0x7f) << (i * 7);

(n[i] & 0x7f) - will be an int value (32bit in most cases) and if
(i * 7) is larger than 32, we will shift outside of 32bit and 0
will be written into *out which is not what we want. To prevent
this, left operand of << should be first casted to 64bit and then
shift should be performed.

Marking no_ci since this code is not used in embedlog, and is
provided as an example for decoding numbers in binary logs.

no_ci
Reported-by: pvs-studio
Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>